fix: window changes its size when its resizable state is changed
When after window resizable state is changed from true to false
and it's forced to be redrawn (by minimizing/restoring or moving
it) its size is shrinken by 10 pixels in hight and width. It happens
because we first set minimum and maximum size of the window to
the current size and after that we make a call to SetCanResize
which removes WS_THICKFRAME from the window's style. After that
when window is moved (or minimized and restored) WM_GETMINMAXINFO
is generated and it bases on min/max values which we set earlier.
The problem is that we set values valid for window with different
style (removing WS_THICKFRAME removes sizing box placed around the
window which is 5 pixels wide) and not for the current one and it
results in window being shrinked. This commit fixes it by changing
order of setting size constraints and setting window's style via
call to SetCanResize.